Add payments page
Added Payments Page but unable to check if it integrates well due to requiring sign in.
Currently the view of the main payments page is almost done.
Dummy data is used as the API of the payments is not yet completed.
